+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+## [2.0.0] - 2023-10-27
+
 ### Fixed
 
 - Allow viewing the version from inside any directory
@@ -121,7 +123,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/spapanik/eulertools/compare/v1.0.0...main
+[Unreleased]: https://github.com/spapanik/eulertools/compare/v2.0.0...main
+[2.0.0]: https://github.com/spapanik/yamk/compare/v1.0.0...v2.0.0
 [1.0.0]: https://github.com/spapanik/yamk/compare/v0.10.0...v1.0.0
 [0.10.0]: https://github.com/spapanik/yamk/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/spapanik/yamk/compare/v0.8.0...v0.9.0